configure with severity and description
